Task-Url: https://github.com/GovernIB/ripea/issues/66

+++ distribucio-core/src/main/java/es/caib/distribucio/core/service/BustiaServiceImpl.java
@@ -17,6 +17,7 @@ import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
 import javax.mail.internet.MimeMessage.RecipientType;
 
+import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,6 +30,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import es.caib.distribucio.core.helper.PropertiesHelper;
 
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
@@ -98,6 +100,13 @@ import es.caib.distribucio.core.repository.ReglaRepository;
 import es.caib.distribucio.core.repository.UnitatOrganitzativaRepository;
 import es.caib.distribucio.core.security.ExtendedPermission;
 
+import java.util.concurrent.TimeUnit;
+import org.springframework.beans.factory.annotation.Autowired;
+import com.codahale.metrics.MetricRegistry;
+import com.codahale.metrics.Timer;
+import com.codahale.metrics.json.MetricsModule;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 
 /**
  * Implementació dels mètodes per a gestionar bústies.
@@ -163,7 +172,6 @@ public class BustiaServiceImpl implements BustiaService {
 	
 	@Autowired
 	private MetricRegistry metricRegistry;
-	  
 
 
 	@Override
@@ -352,6 +360,18 @@ public class BustiaServiceImpl implements BustiaService {
 				}			
 			}
 		}
+		
+		// cannot remove busties containing any anotacions
+		if (contingutRepository.findByPare(bustia) != null && !contingutRepository.findByPare(bustia).isEmpty()) {
+			String missatgeError = "No es pot esborrar la bústia amb anotacions a dins (" + 
+					"bustiaId=" + id + ")";
+			logger.error(missatgeError);
+			throw new ValidationException(
+					id,
+					BustiaEntity.class,
+					missatgeError);
+		}
+		
 		bustiaRepository.delete(bustia);
 		return toBustiaDto(
 				bustia,
@@ -629,6 +649,10 @@ public class BustiaServiceImpl implements BustiaService {
 	@Transactional(readOnly = true)
 	public List<BustiaDto> findPermesesPerUsuari(
 			Long entitatId) {
+		
+		final Timer findPermesesPerUsuariTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "findPermesesPerUsuari"));
+		Timer.Context findPermesesPerUsuariContext = findPermesesPerUsuariTimer.time();
+		
 		logger.debug("Consulta de busties permeses per un usuari ("
 				+ "entitatId=" + entitatId + ")");
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
@@ -651,7 +675,11 @@ public class BustiaServiceImpl implements BustiaService {
 				BustiaEntity.class,
 				new Permission[] {ExtendedPermission.READ},
 				auth);
-		List<BustiaDto> bustiesRetorn = toBustiaDto(busties, true, true);
+		List<BustiaDto> bustiesRetorn = toBustiaDto(busties, false, true);
+		
+		
+		findPermesesPerUsuariContext.stop();
+		
 		return bustiesRetorn;
 	}
 
@@ -818,8 +846,7 @@ public class BustiaServiceImpl implements BustiaService {
 			Long entitatId,
 			Long contingutId,
 			Long registreId, 
-			String adresses,
-			String serverPortContext) throws MessagingException {
+			String adresses) throws MessagingException {
 		SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy HH:mm:ss");
 		RegistreAnotacioDto registre = registreService.findOne(
 				entitatId,
@@ -848,6 +875,9 @@ public class BustiaServiceImpl implements BustiaService {
 				entitatId,
 				contingutId,
 				registreId);
+		
+		String appBaseUrl = PropertiesHelper.getProperties().getProperty("es.caib.distribucio.app.base.url");
+		
 		MimeMessage missatge = mailSender.createMimeMessage();
 		missatge.setHeader("Content-Type", "text/html charset=UTF-8");
 		MimeMessageHelper helper;
@@ -908,7 +938,7 @@ public class BustiaServiceImpl implements BustiaService {
 							"				<th>"+ messageHelper.getMessage("registre.annex.detalls.camp.fitxer") + "</th>"+
 							"				<td>"  + Objects.toString(justificant.getFitxerNom(), "") + "("+Objects.toString(justificant.getFitxerTamany(), "")+" bytes)"+
 							
-							"<a href=\"http://"+serverPortContext+"/modal/contingut/"+contingutId+"/registre/"+registreId+"/justificant\"> Descarregar </a>" +
+							"<a href=\""+appBaseUrl+"/modal/contingut/"+contingutId+"/registre/"+registreId+"/justificant\"> Descarregar </a>" +
 							
 							"</td>"+
 							"			</tr>"+								
@@ -1032,7 +1062,7 @@ public class BustiaServiceImpl implements BustiaService {
 					"				<th>"+ messageHelper.getMessage("registre.annex.detalls.camp.fitxer") + "</th>"+
 					"				<td>"  + Objects.toString(annex.getFitxerNom(), "") + "("+Objects.toString(annex.getFitxerTamany(), "")+" bytes)"+
 					
-					"<a href=\"http://"+serverPortContext+"/modal/contingut/"+contingutId+"/registre/"+registreId+"/annex/"+annex.getId()+"/arxiu/DOCUMENT\"> Descarregar </a>"
+					"<a href=\""+appBaseUrl+"/modal/contingut/"+contingutId+"/registre/"+registreId+"/annex/"+annex.getId()+"/arxiu/DOCUMENT\"> Descarregar </a>"
 					+
 					"</td>"+
 					"			</tr>"+
@@ -1674,14 +1704,10 @@ public class BustiaServiceImpl implements BustiaService {
 				+ "paginacioParams=" + paginacioParams + ")");
 		
 		
-		
-		final Timer timerTotal = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable"));
+		final Timer timerTotal = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "contingutPendentFindByDatatable"));
 		Timer.Context contextTotal = timerTotal.time();
-
-		
-		
-		
-		final Timer comprovarEntitatTimer = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable.comprovarEntitat"));
+	
+		final Timer comprovarEntitatTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "contingutPendentFindByDatatable.comprovarEntitat"));
 		Timer.Context comprovarEntitatContext = comprovarEntitatTimer.time();
 		
 		EntitatEntity entitat = entityComprovarHelper.comprovarEntitat(
@@ -1691,10 +1717,8 @@ public class BustiaServiceImpl implements BustiaService {
 				false);
 		
 		comprovarEntitatContext.stop();
-		
-
-		
-		final Timer comprovarBustiaTimer = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable.comprovarBustia"));
+			
+		final Timer comprovarBustiaTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "contingutPendentFindByDatatable.comprovarBustia"));
 		Timer.Context comprovarBustiaContext = comprovarBustiaTimer.time();
 		
 		// Comprova la bústia i que l'usuari hi tengui accés
@@ -1720,8 +1744,7 @@ public class BustiaServiceImpl implements BustiaService {
 		comprovarBustiaContext.stop();
 		
 		
-		
-		final Timer findRegistreByPareAndFiltreTimer = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable.findRegistreByPareAndFiltre"));
+		final Timer findRegistreByPareAndFiltreTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "contingutPendentFindByDatatable.findRegistreByPareAndFiltre"));
 		Timer.Context findRegistreByPareAndFiltreContext = findRegistreByPareAndFiltreTimer.time();
 		
 		Map<String, String[]> mapeigOrdenacio = new HashMap<String, String[]>();
@@ -1762,7 +1785,7 @@ public class BustiaServiceImpl implements BustiaService {
 					(filtre.getDataRecepcioInici() == null),
 					filtre.getDataRecepcioInici(),
 					(filtre.getDataRecepcioFi() == null),
-					filtre.getDataRecepcioFi(),
+					new DateTime(filtre.getDataRecepcioFi()).plusDays(1).toDate(), 
 					registreEstat == null,
 					registreEstat,
 					paginacioParams.getFiltre() == null || paginacioParams.getFiltre().isEmpty(),
@@ -1776,7 +1799,7 @@ public class BustiaServiceImpl implements BustiaService {
 		findRegistreByPareAndFiltreContext.stop();
 		
 		
-		final Timer toPaginaDtoTimer = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable.toPaginaDto"));
+		final Timer toPaginaDtoTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "contingutPendentFindByDatatable.toPaginaDto"));
 		Timer.Context toPaginaDtoContext = toPaginaDtoTimer.time();
 		
 		PaginaDto<BustiaContingutDto> pag = paginacioHelper.toPaginaDto(
@@ -2126,7 +2149,7 @@ public class BustiaServiceImpl implements BustiaService {
 			ContingutEntity contingut) {
 		
 		// TIMER START
-		final Timer getPathContingutComDtoTimer = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable.toPaginaDto.toBustiaContingutDto.getPathContingutComDto"));
+		final Timer getPathContingutComDtoTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "toBustiaContingutDto.getPathContingutComDto"));
 		Timer.Context getPathContingutComDtoContext = getPathContingutComDtoTimer.time();
 		
 		
@@ -2140,15 +2163,12 @@ public class BustiaServiceImpl implements BustiaService {
 				false);
 		bustiaContingut.setPath(path);
 		
-		
 		getPathContingutComDtoContext.stop();
 		// TIMER STOP
-		
-		
-		
+			
 		
 		// TIMER START
-		final Timer toBustiaDtoTimer = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable.toPaginaDto.toBustiaContingutDto.toBustiaDto"));
+		final Timer toBustiaDtoTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "toBustiaContingutDto.toBustiaDto"));
 		Timer.Context toBustiaDtoContext = toBustiaDtoTimer.time();
 		
 		BustiaDto pare = toBustiaDto((BustiaEntity)(contingut.getPare()), false, false);
@@ -2166,7 +2186,7 @@ public class BustiaServiceImpl implements BustiaService {
 		
 		
 		// TIMER START
-		final Timer countByContingutTimer = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable.toPaginaDto.toBustiaContingutDto.countByContingut"));
+		final Timer countByContingutTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "toBustiaContingutDto.countByContingut"));
 		Timer.Context countByContingutContext = countByContingutTimer.time();
 		
 		RegistreEntity registre = null;
@@ -2205,7 +2225,7 @@ public class BustiaServiceImpl implements BustiaService {
 		
 		
 		// TIMER START
-		final Timer countByLlegidaAndContingutIdTimer = metricRegistry.timer(MetricRegistry.name(BustiaService.class, "contingutPendentFindByDatatable.toPaginaDto.toBustiaContingutDto.countByLlegidaAndContingutId"));
+		final Timer countByLlegidaAndContingutIdTimer = metricRegistry.timer(MetricRegistry.name(BustiaServiceImpl.class, "toBustiaContingutDto.countByLlegidaAndContingutId"));
 		Timer.Context countByLlegidaAndContingutIdContext = countByLlegidaAndContingutIdTimer.time();
 		
 		bustiaContingut.setAlerta(alertaRepository.countByLlegidaAndContingutId(
